requirements.txt
slackclient==2.5.0
requests
sanic
git+https://github.com/chauffer/cobe@master
